requirements.txt
pypdf>=4.0.0
openai>=1.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
langgraph-cli>=0.0.40

